EazySkool/requirements.txt
bcrypt==4.0.1
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.2.0 